+++ src/SoftTissue.Core/Operations/FileManager/SkipPoint/ISkipPoint.cs
@@ -1,6 +1,0 @@
-﻿namespace SoftTissue.Core.Operations.FileManager.SkipPoint
-{
-    public interface ISkipPoint
-    {
-    }
-}