+++ src/SoftTissue.Core/Operations/FileManager/SkipPoint/SkipPoint.cs
@@ -1,6 +1,0 @@
-﻿namespace SoftTissue.Core.Operations.FileManager.SkipPoint
-{
-    public class SkipPoint : ISkipPoint
-    {
-    }
-}