+++ src/SoftTissue.DataContract/FileManager/SkipPoints/SkipPointsResponse.cs
@@ -1,0 +1,9 @@
+﻿using SoftTissue.DataContract.OperationBase;
+
+namespace SoftTissue.DataContract.FileManager.SkipPoints
+{
+    /// <summary>
+    /// It represents the response content for SkipPoints operations.
+    /// </summary>
+    public class SkipPointsResponse : OperationResponseBase<SkipPointsResponseData> { }
+}