+++ etc/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## UNRELEASED
 ### Added
 - Data contracts to CalculateResult operation.
+- Operation SkipPoints.
 - Operation AnalyzeAndPredictResults.
 - Operation GenerateDomain.
 - Operation CalculateConvergenceTime.